Update data_policy.md
I tried fine tuning the wording, I am not sure it is better, but have a look.  In the end I changed my mind and thought that a listing of the overarching institutions supporting/funding the project would probably save us lots of hassle later on, but I ordered these a bit by my sense of their initiating role.   Please feel free to nit pick