First iteration on flight report template
I tried to adapt the flight report template to match my expectations and am struggling a bit on how to handle the formatting.  For somethings we might want person oriented information, i.e. what was their job, for other things we might want instrument specific information, i.e., did it work okay, who was responsible for it.

For instrument snapshots I think it would be good to have some standard picture that are repeated for each flight.  However we should be able to add some instrument impressions and to the impression section others could add their impressions.

I'd appreciate if others look this over and comment.  Having a quick first merge and continuing to update it via issues and PRs would be my preferred way of working if we want to get this in place in time for others to adopt the workflow