Update flight plan forecasts and cleanup
In addition to updating the valid forecast times the flightplan
was restructured.  The restructing separates the creation from
the flight plan from its presentation.   It also tries to minimize
the code and maximize readibility.

Some of the structuring anticipates a snap-to-satellite track
functionality, as well as more condensed output functionality.