restrict zarr to version 2
Zarr 3 has been released. For sure, we want to support zarr 3, the
sooner the better. However, there still seem to be some teething
problems with respect to backwards compatibility. So for now, it seems
to be [best
practice](https://zarr.readthedocs.io/en/stable/user-guide/v3_migration.html#getting-ready-for-3-0)
to stick with zarr 2 for the moment.